requirements.txt
torch>=1.12
numpy>=1.22
matplotlib>=3.5
seaborn>=0.11
Pillow>=9.0
pytest>=7.0